app/requirements.txt
asgiref==3.2.10
Django==3.0.7
gunicorn==20.0.4
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
channels>=2.1.5
channels-redis>=2.3.1
daphne
djangorestframework==3.12.1
django-cors-headers
boto3==1.15.16
botocore==1.18.16
jmespath==0.10.0
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.10
requests==2.24.0
cryptography~=2.8
drf-jwt~=1.13.3
pyjwt~=1.7.1
